Add confirm, sorting by creation time, filtering by text, also test the loc.geo.address, grouping by lastUpdated.

+++ js/app.controller.js
@@ -83,16 +83,19 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
-    locService.remove(locId)
-        .then(() => {
-            flashMsg('Location removed')
-            unDisplayLoc()
-            loadAndRenderLocs()
-        })
-        .catch(err => {
-            console.error('OOPs:', err)
-            flashMsg('Cannot remove location')
-        })
+    locService.confirmRemove().then((confirmed) => {
+    if (confirmed) {
+        locService.remove(locId)
+            .then(() => {
+                flashMsg('Location removed')
+                unDisplayLoc()
+                loadAndRenderLocs()
+            })
+            .catch(err => {
+                console.error('OOPs:', err)
+                flashMsg('Cannot remove location')
+            })
+  } })  
 }
 
 function onSearchAddress(ev) {
@@ -280,6 +283,9 @@ function onSetFilterBy({ txt, minRate }) {
 function renderLocStats() {
     locService.getLocCountByRateMap().then(stats => {
         handleStats(stats, 'loc-stats-rate')
+    })
+    locService.getLocCountByLastUpdated().then(stats => {
+        handleStats(stats, 'loc-stats-lastUpdated')
     })
 }
